+++ demos/spring-boot-starter-demo/src/main/java/org/rockyang/mybatis/boot/demo/controller/HelloController.java
@@ -24,6 +24,11 @@ public class HelloController {
 		return "Hello Mybatis-Kits.";
 	}
 
+	@GetMapping("/newId")
+	public Long newId() {
+		return userMapper.getNewId();
+	}
+
 	/**
 	 * 获取用户列表
 	 * @return