+++ mybatis-kits-spring-boot/mybatis-kits-spring-boot-autoconfigure/src/main/java/org/rockyang/mybatis/boot/autoconfigure/MybatisProperties.java
@@ -21,10 +21,8 @@ import java.util.Properties;
  * @author chenzhaoju
  * @author yangjian
  */
-@ConfigurationProperties(prefix = MybatisProperties.MYBATIS_PREFIX)
+@ConfigurationProperties(prefix = "mybatis")
 public class MybatisProperties {
-
-  public static final String MYBATIS_PREFIX = "mybatis";
 
   /**
    * Location of MyBatis xml config file.