+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 # Changelog
 
 -------------------------------------------------------------------------------------------------------------
+# 1.5.2
+* 完成文档编写
+* 添加 Twitter Snowflake 分布式 ID 生成算法实现
+* 删除 IdUtil 和 IdUtils ID 生成工具，使用 Snowflake 替换
+
+# 1.5.1
+* 修复分页插件注入失效 Bug
+* 修改 mybatis-spring-boot 构件名称为 mybatis-kits-spring-boot-starter
+* 更新单元测试以及配置文档
 
 # 1.5.0
 * 更新包名称为 org.rockyang